+++ azure_img_utils/cli/cli_utils.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+
+"""Azure image utils cli module utils."""
+
+# Copyright (c) 2022 SUSE LLC
+#
+# This file is part of azure_img_utils. azure_img_utils provides an
+# api and command line utilities for handling images in the Azure Cloud.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import click
+import logging
+
+# -----------------------------------------------------------------------------
+# Shared options
+shared_options = [
+    click.option(
+        '--no-color',
+        is_flag=True,
+        help='Remove ANSI color and styling from output.'
+    ),
+    click.option(
+        '--verbose',
+        'log_level',
+        flag_value=logging.DEBUG,
+        help='Display debug level logging to console.'
+    ),
+    click.option(
+        '--info',
+        'log_level',
+        flag_value=logging.INFO,
+        default=True,
+        help='Display logging info to console. (Default)'
+    ),
+    click.option(
+        '--quiet',
+        'log_level',
+        flag_value=logging.ERROR,
+        help='Display only errors to console.'
+    )
+]
+
+
+def add_options(options):
+    def _add_options(func):
+        for option in reversed(options):
+            func = option(func)
+        return func
+    return _add_options
+
+
+# -----------------------------------------------------------------------------
+# Printing options
+def echo_style(message, no_color, fg='yellow'):
+    """
+    Echo stylized output to terminal depending on no_color.
+    """
+    if no_color:
+        click.echo(message)
+    else:
+        click.secho(message, fg=fg)
